Which Oxford? Will fix #21
The text in the description of the Life-long, recipient-managed Credentials (education) focal use case is full of jargon and assumptions about the solution. However, we have stated (in section 1.1) that we're not starting from scratch so this is, just about, tolerable. However, I have added a hyperlink to the University of Oxford and made it clear that it's important that it's that specific university when the credential is verified 10 years later.

Meanwhile, I can't help but chortle at two Americans discussing the likelihood of the University of Oxford still existing in 50 years' time. America ca. 1776, Uni Oxford ca. 1096.